[arithmetic] Organize `Subtract` (#964)
