+++ qualtran/bloqs/arithmetic/_shims.py
@@ -41,15 +41,6 @@ class MultiCToffoli(Bloq):
 
 
 @frozen
-class Sub(Bloq):
-    n: int
-
-    @cached_property
-    def signature(self) -> 'Signature':
-        return Signature([Register('x', QUInt(self.n)), Register('y', QUInt(self.n))])
-
-
-@frozen
 class Lt(Bloq):
     n: int
     signed: bool = False
